+++ controllers/contacts/listContacts.js
@@ -1,0 +1,9 @@
+const contactsOperations = require('../../model/contacts')
+const { sendSuccessRes } = require('../../helpers')
+
+const listContacts = async (req, res) => {
+  const result = await contactsOperations.listContacts()
+  sendSuccessRes(res, { result })
+}
+
+module.exports = listContacts
