+++ helpers/httpErrors.js
@@ -1,9 +1,14 @@
-const { NotFound } = require('http-errors')
+const { NotFound, BadRequest } = require('http-errors')
 
 const notFound = (contactId, next) => {
   next(new NotFound(`object with id=${contactId} not found`))
 }
 
+const badRequest = (error, next) => {
+  next(new BadRequest(error.message))
+}
+
 module.exports = {
   notFound,
+  badRequest,
 }
