+++ middlewares/validation.js
@@ -1,15 +1,11 @@
+const { badRequest } = require('../helpers')
+
 const validation = (schema) => {
-  return async (req, res, next) => {
+  return async (req, _, next) => {
     const { error } = schema.validate(req.body)
 
     if (error) {
-      res.status(400).json({
-        status: 'error',
-        code: 400,
-        message: error.message,
-      })
-
-      return
+      return badRequest(error, next)
     }
     next()
   }
