+++ model/contacts/getContactById.js
@@ -2,7 +2,7 @@ const listContacts = require('./listContacts')
 
 const getContactById = async (contactId) => {
   const contacts = await listContacts()
-  const contact = contacts.find((item) => item.id.toString() === contactId)
+  const contact = contacts.find((item) => item.id === contactId)
 
   if (!contact) {
     return null
