--- routes/api/contacts/index.js
@@ -1,3 +1,0 @@
-const router = require('./contacts')
-
-module.exports = router
